# backend/openapi/openapi.json
{
  "components": {
    "schemas": {
...
      "DupGroupOut": {
        "properties": {
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "members": {
            "items": {
              "$ref": "#/components/schemas/DupMemberOut"
            },
            "title": "Members",
            "type": "array"
          },
          "primary_document_id": {
            "title": "Primary Document Id",
            "type": "integer"
          },
          "resolved": {
            "title": "Resolved",
            "type": "boolean"
          },
          "similarity_type": {
            "title": "Similarity Type",
            "type": "string"
          }
        },
        "required": [
          "id",
          "similarity_type",
          "primary_document_id",
          "resolved",
          "members"
        ],
        "title": "DupGroupOut",
        "type": "object"
      },
      "DupMemberOut": {
        "properties": {
          "document_id": {
            "title": "Document Id",
            "type": "integer"
          },
          "similarity_score": {
            "title": "Similarity Score",
            "type": "number"
          },
          "title": {
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "document_id",
          "title",
          "similarity_score"
        ],
        "title": "DupMemberOut",
        "type": "object"
      },
      "DuplicatePage": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/DupGroupOut"
            },
            "title": "Items",
            "type": "array"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "required": [
          "items"
        ],
        "title": "DuplicatePage",
        "type": "object"
      },
...
      "ExtractionProvenanceOut": {
        "properties": {
          "completed_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "extractor_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Extractor Name"
          },
          "extractor_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Extractor Version"
          },
          "method": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Method"
          },
          "ocr_engine": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ocr Engine"
          },
          "ocr_engine_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ocr Engine Version"
          },
          "ocr_languages": {
            "items": {
              "type": "string"
            },
            "title": "Ocr Languages",
            "type": "array"
          },
          "quality_score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Quality Score"
          },
          "quality_signals": {
            "additionalProperties": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "number"
                },
                {
                  "type": "integer"
                },
                {
                  "type": "string"
                }
              ]
            },
            "title": "Quality Signals",
            "type": "object"
          }
        },
        "title": "ExtractionProvenanceOut",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "IngestionStatusOut": {
        "properties": {
          "attempt_count": {
            "title": "Attempt Count",
            "type": "integer"
          },
          "cancellable": {
            "title": "Cancellable",
            "type": "boolean"
          },
          "completed_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "degraded_stages": {
            "items": {
              "type": "string"
            },
            "title": "Degraded Stages",
            "type": "array"
          },
          "document_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Document Id"
          },
          "document_status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Document Status"
          },
          "error_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Code"
          },
          "extraction": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ExtractionProvenanceOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "next_attempt_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Attempt At"
          },
          "retryable": {
            "title": "Retryable",
            "type": "boolean"
          },
          "stage": {
            "title": "Stage",
            "type": "string"
          },
          "stage_results": {
            "additionalProperties": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Stage Results",
            "type": "object"
          },
          "stage_version": {
            "title": "Stage Version",
            "type": "string"
          },
          "state": {
            "title": "State",
            "type": "string"
          },
          "terminal": {
            "title": "Terminal",
            "type": "boolean"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          },
          "version_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Version Id"
          }
        },
        "required": [
          "id",
          "state",
          "stage",
          "stage_version",
          "attempt_count",
          "retryable",
          "cancellable",
          "terminal",
          "created_at",
          "updated_at"
        ],
        "title": "IngestionStatusOut",
        "type": "object"
      },
...
      "ResolveDup": {
        "additionalProperties": false,
        "properties": {
          "action": {
            "default": "keep_primary",
            "enum": [
              "keep_primary",
              "keep_both"
            ],
            "title": "Action",
            "type": "string"
          },
          "primary_document_id": {
            "exclusiveMinimum": 0.0,
            "title": "Primary Document Id",
            "type": "integer"
          }
        },
        "required": [
          "primary_document_id"
        ],
        "title": "ResolveDup",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/v1/auth/login"
          }
        },
        "type": "oauth2"
      }
    }
  },
  "info": {
    "description": "AI-based Digital Document Management System — local vertical slice.",
    "title": "XENIUS DocVault API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/duplicates": {
      "get": {
        "operationId": "list_duplicate_groups_api_v1_duplicates_get",
        "parameters": [
          {
            "in": "query",
            "name": "include_resolved",
            "required": false,
            "schema": {
              "default": false,
              "title": "Include Resolved",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/DupGroupOut"
                  },
                  "title": "Response List Duplicate Groups Api V1 Duplicates Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "List Duplicate Groups",
        "tags": [
          "duplicates"
        ]
      }
    },
    "/api/v1/duplicates/cursor": {
      "get": {
        "operationId": "list_duplicate_groups_cursor_api_v1_duplicates_cursor_get",
        "parameters": [
          {
            "in": "query",
            "name": "include_resolved",
            "required": false,
            "schema": {
              "default": false,
              "title": "Include Resolved",
              "type": "boolean"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 256,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "maximum": 500,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DuplicatePage"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "List Duplicate Groups Cursor",
        "tags": [
          "duplicates"
        ]
      }
    },
    "/api/v1/duplicates/{group_id}/resolve": {
      "post": {
        "operationId": "resolve_group_api_v1_duplicates__group_id__resolve_post",
        "parameters": [
          {
            "in": "path",
            "name": "group_id",
            "required": true,
            "schema": {
              "minimum": 1,
              "title": "Group Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResolveDup"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DupGroupOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Resolve Group",
        "tags": [
          "duplicates"
        ]
      }
    },
...
    "/api/v1/ingestions/{job_id}": {
      "get": {
        "operationId": "get_ingestion_status_api_v1_ingestions__job_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "maxLength": 36,
              "minLength": 1,
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IngestionStatusOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Get Ingestion Status",
        "tags": [
          "ingestions"
        ]
      }
    },
    "/api/v1/ingestions/{job_id}/cancel": {
      "post": {
        "operationId": "cancel_ingestion_api_v1_ingestions__job_id__cancel_post",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "maxLength": 36,
              "minLength": 1,
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IngestionStatusOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Cancel Ingestion",
        "tags": [
          "ingestions"
        ]
      }
    },
    "/api/v1/ingestions/{job_id}/retry": {
      "post": {
        "operationId": "retry_ingestion_api_v1_ingestions__job_id__retry_post",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "maxLength": 36,
              "minLength": 1,
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IngestionStatusOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Retry Ingestion",
        "tags": [
          "ingestions"
        ]
      }
    },
...
  }
}